+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Add Support for Java 26 and 27
+- Show preview class file versions in the structure view
 
 ### Changed